edit question works

+++ src/store/technologies/actions.ts
@@ -5,10 +5,11 @@ import {AnyAction, ThunkAction} from "@reduxjs/toolkit";
 import {TechnologyDataPayload} from "../../api/model/technologyDataPayload";
 import {addFlag} from "../shared/page/page.slice";
 import {errorFlag, successFlag} from "../../service/flag.service";
-import {CreateQuestionForm} from "../../components/pages/technologies/addQuestion/AddQuestionForm";
 import {QuestionService} from "../../service/question.service";
 import {Technology} from "../../api/model/technology.model";
 import {isNil} from "lodash-es";
+import {QuestionFormModel} from "../../components/pages/technologies/questionForm/QuestionFormModel";
+import QuestionRepository from "../../api/repository/questions.repository";
 
 export const fetchTechnologies = () => async (dispatch: Dispatch) => {
     try {
@@ -47,12 +48,12 @@ export const addTechnology = (values: TechnologyDataPayload): ThunkAction<Promis
 
 export const editTechnology = (values: TechnologyDataPayload): ThunkAction<Promise<boolean>, RootState, unknown, AnyAction> => async (dispatch, getState) => {
     try {
-        const technologyId = selectTechnologyContextId(getState());
+        const technologyId = selectTechnologyContextId(getState()) || 0;
 
         const {data} = await TechnologiesRepository.editTechnology(technologyId, values)
 
         dispatch(setTechnology(data))
-        dispatch(addFlag(successFlag("Technology edited successfully.")))
+        dispatch(addFlag(successFlag("Technologia edytowana poprawnie")))
         return true;
     } catch (e) {
         console.log(e)
@@ -61,17 +62,47 @@ export const editTechnology = (values: TechnologyDataPayload): ThunkAction<Promi
     }
 }
 
-export const addQuestion = (values: CreateQuestionForm): ThunkAction<Promise<boolean>, RootState, unknown, AnyAction> => async (dispatch, getState) => {
-    const technologyId = selectTechnologyContextId(getState());
+export const addQuestion = (values: QuestionFormModel): ThunkAction<Promise<boolean>, RootState, unknown, AnyAction> => async (dispatch, getState) => {
+    const technologyId = selectTechnologyContextId(getState()) || 0;
 
     try {
         await QuestionService.createQuestion(values, technologyId)
-        dispatch(addFlag(successFlag("Question saved successfully.")))
+        dispatch(addFlag(successFlag("Pytanie dodane poprawnie")))
         dispatch(fetchTechnology(technologyId))
         return true;
     } catch (e) {
         console.log(e)
-        dispatch(addFlag(errorFlag("Cannot add question")))
+        dispatch(addFlag(errorFlag("Nie dodano pytania")))
+        return false;
+    }
+}
+
+export const editQuestion = (questionId: number, values: QuestionFormModel): ThunkAction<Promise<boolean>, RootState, unknown, AnyAction> => async (dispatch, getState) => {
+    const technologyId = selectTechnologyContextId(getState()) || 0;
+
+    try {
+        await QuestionService.editQuestion(questionId, values)
+        dispatch(addFlag(successFlag("Pytanie zapisane poprawnie")))
+        dispatch(fetchTechnology(technologyId))
+        return true;
+    } catch (e) {
+        console.log(e)
+        dispatch(addFlag(errorFlag("Nie edytowano pytania")))
+        return false;
+    }
+}
+
+export const deleteQuestion = (questionId: number): ThunkAction<Promise<boolean>, RootState, unknown, AnyAction> => async (dispatch, getState) => {
+    const technologyId = selectTechnologyContextId(getState()) || 0;
+
+    try {
+        await QuestionRepository.deleteQuestion(questionId)
+        dispatch(addFlag(successFlag("Pytanie usunięte poprawnie")))
+        dispatch(fetchTechnology(technologyId))
+        return true;
+    } catch (e) {
+        console.log(e)
+        dispatch(addFlag(errorFlag("Nie usunięto pytania")))
         return false;
     }
 }
@@ -79,6 +110,11 @@ export const addQuestion = (values: CreateQuestionForm): ThunkAction<Promise<boo
 
 export const deleteTechnology = (): ThunkAction<void, RootState, unknown, AnyAction> => async (dispatch, getState) => {
     const technologyId = selectTechnologyContextId(getState())
+
+    if (isNil(technologyId)) {
+        console.error("Not found technology context id")
+        return;
+    }
     try {
         await TechnologiesRepository.remove(technologyId)
     } catch (e) {
